scripts/ab-test/pyproject.toml
[project]
name = "superpowers-beads-ab-test"
version = "0.1.0"
description = "A/B test harness measuring skill-description activation rates against fresh Claude sessions."
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6",
]

[project.optional-dependencies]
test = []
